requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.0
pydantic==2.5.0
python-multipart==0.0.6
redis==5.0.1
httpx<0.25.0,>=0.24.0
python-dotenv==1.0.0
asyncpg==0.29.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
faker==20.1.0

# JWT and OAuth dependencies
PyJWT==2.8.0
cryptography==41.0.7
